==> src/03a_Prepare_biodiversity_data.R <==
#- - - - - - - - - - - - - - - - -#
#   Data preparation for SDMs     #
#                                 #
#     author: Romy Zeiss          #
#       date: 18.08.2021          #
#- - - - - - - - - - - - - - - - -#

#setwd("T:/_students/Romy/SoilBiodiversity")

gc()
library(tidyverse)
library(here)

library(CoordinateCleaner)


#- - - - - - - - - - - - - - - - - - - - -
Taxon_name <- "Crassiclitellata"
speciesNames <- read.csv(file=paste0("./results/Species_list_", Taxon_name, ".csv"))

# geographic extent of Europe
extent_Europe <- c(-23, 60, 31, 75)

# - - - - - - - - - - - - - - - - - - -
## Data from sWorm ####
sworm_occ <- readr::read_csv(file=paste0(here::here(), "/data_raw/SppOccData_sWorm_v2.csv"))
sworm_site <- readr::read_csv(file=paste0(here::here(), "/data_raw/SiteData_sWorm_v2.csv"))

sworm <- dplyr::full_join(sworm_occ, sworm_site)

# extract year (of publication)
sworm$year <- stringr::str_extract(sworm$ID,"[:alpha:]+[:digit:]{4}")
sworm$year <- stringr::str_extract(sworm$year,"[:digit:]{4}")
sworm[sworm$year > 2020 & !is.na(sworm$year), "year"] <- NA
sworm[sworm$year < 1900 & !is.na(sworm$year), "year"] <- NA #note: last real dates in 1986
sworm$year <- as.numeric(sworm$year)

summary(sworm$year)

sworm <- sworm %>% filter(Abundance>0, !is.na(SpeciesBinomial),
                          year >= 1970) %>%
  dplyr::select(SpeciesBinomial, Latitude_decimal_degrees, Longitude_decimal_degrees, year)

sworm$datasource <- "sWorm"

rm(sworm_occ, sworm_site)

# - - - - - - - - - - - - - - - - - - -
## Data from GBIF ####

dat <- read.csv(file=paste0(here::here(), "/intermediates/Occurrences_GBIF_Crassiclitellata.csv")) #dat

dat <- dat %>% filter(occurrenceStatus == "PRESENT") # remove absent records

gbif <- dat[,c("species", "decimalLatitude", "decimalLongitude", "year")] %>%
  mutate(datasource = "GBIF") 

rm(dat)

# - - - - - - - - - - - - - - - - - - -
## Data from Edaphobase ####

edapho <- readr::read_csv(file=paste0(here::here(), "/data_raw/Edaphobase_download_24-Feb-2021_Lumbricidae_Europe.csv"))

#!!! Manually: We already added one missing "Valid taxon" for Helodrilus sp.

# extract the first 2 words of the species' names
edapho$species <- word(edapho$"Valid taxon", 1, 2)
edapho$species <- sub(",*", "\\1", edapho$species)
edapho$species <- gsub(",", "", edapho$species)

# exclude species with only Genus name (followed by identifier's name)
edapho <- edapho %>% filter(!str_detect(species, " [:upper:]"))

# remove observations before 1970
edapho$date <- as.Date(edapho$"Observation date (Sampling event)", format="%m/%d/%y") 
edapho <- edapho %>% filter(date >= "1970-01-01") %>% mutate(year=substr(date, 1,4))
edapho$year <- as.numeric(edapho$year)

edapho$datasource <- "Edaphobase"

# have a look at the data
edapho <- edapho[,c("species", "Latitude", "Longitude", "year", "datasource")] #n=13.473
edapho

# - - - - - - - - - - - - - - - - - - -
## Data from SoilReCon project (Portugal) ####
  
recon <- readr::read_csv(file=paste0(here::here(), "/data_raw/SoilReCon_earthworms_clean.csv"))

# remove non-species level and NA species
recon <- recon[recon$Species!="Octolasion sp." & !is.na(recon$Species),]

# filter relevant columns
recon <- recon %>% dplyr::select(Species, POINT_X, POINT_Y) %>%
  mutate("year"=2021, "datasource"="SoilReCon")

# - - - - - - - - - - - - - - - - - - -
## Data from Jerome Matthieu (Jema) ####

jema <-  read.csv(file=paste0(here::here(),"/data_raw/worm_spd_europe_jerome.csv"), sep=";")
# nrow=27800

jema$Species <- jema$species_name

# rename "A. caliginosa trapezoides"
jema[jema$Species=="A. caliginosa trapezoides","Species"] <- "Aporrectodea caliginosa"

# make upper letter
jema$Species <- stringr::str_to_sentence(jema$Species)

# remove underscore in species names
jema$Species <- gsub("_", " ", jema$Species)

# keep only species, not subspecies
jema$Species <- word(jema$Species, 1, 2)

# remove - from species names
jema$Species <- gsub("[[:punct:]]", "",jema$Species)

# remove sp species
jema <- jema %>% filter(jema$Species %in% jema$Species[!stringr::str_detect(jema$Species, "sp[p]?[:blank:]")])

jema$lat <- as.double(stringr::str_replace(jema$lat, ",", "."))
jema$lon <- as.double(stringr::str_replace(jema$lon, ",", "."))

# plot raw records
world.inp <- map_data("world")

pdf(paste0(here::here(), "/figures/OccurrenceRaw_JM_perYear.pdf"), width=10)
ggplot()+
  geom_map(data = world.inp, map = world.inp, aes(map_id = region), fill = "white")+
  geom_point(data=jema, aes(x=lon, y=lat, col=obs_year), cex=0.3)+ theme_bw()+
  xlim(min(extent_Europe[1], na.rm = T), max(extent_Europe[2], na.rm = T)) +
  ylim(min(extent_Europe[3], na.rm = T), max(extent_Europe[4], na.rm = T)) +
  scale_color_steps2(breaks=c(1900, 1950, 1970, 1990), midpoint=1960, 
                     high="#10a53dFF", mid="#ffcf20FF", low="#541352FF")+
  theme(panel.background = element_rect(fill = "grey80",
                                colour = "grey80",
                                size = 0.5, linetype = "solid"),
        panel.grid.minor = element_line(size = 0.5, linetype = 'solid',
                                        colour = "grey80"))
dev.off()

# remove data before 1970
jema <- jema %>% filter(obs_year >=1970)

jema <- jema %>% rename("datasource"=source) %>% dplyr::select(Species, lat, lon, obs_year, datasource)
# nrow=26414 (1990: 18166 records)

# - - - - - - - - - - - - - - - - - - -
## Merge all together ####

data_raw <- dplyr::full_join(sworm, gbif, 
                         by=c("SpeciesBinomial"="species", 
                              "Latitude_decimal_degrees"="decimalLatitude", 
                              "Longitude_decimal_degrees"="decimalLongitude", 
                              "datasource", "year"))

data_raw <- data_raw %>% dplyr::full_join(edapho, 
                 by=c("SpeciesBinomial"="species", 
                      "Latitude_decimal_degrees"="Latitude", 
                      "Longitude_decimal_degrees"="Longitude",
                      "datasource", "year"))

data_raw <- data_raw %>% dplyr::full_join(recon, 
                         by=c("SpeciesBinomial"="Species", 
                              "Latitude_decimal_degrees"="POINT_Y", 
                              "Longitude_decimal_degrees"="POINT_X",
                              "datasource", "year"))

data_raw <- data_raw %>% dplyr::full_join(jema, 
                         by=c("SpeciesBinomial"="Species", 
                              "Latitude_decimal_degrees"="lat", 
                              "Longitude_decimal_degrees"="lon",
                              "datasource", "year"="obs_year"))

data_raw <- tibble::tibble(species=data_raw$SpeciesBinomial, 
                       latitude=data_raw$Latitude_decimal_degrees, 
                       longitude=data_raw$Longitude_decimal_degrees, 
                       datasource=data_raw$datasource, 
                       year=data_raw$year)

data_raw #nrow = 108,027 (1990: 101,190)

# - - - - - - - - - - - - - - - - - - -
## Save data ####
write.csv(data_raw, file=paste0(here::here(), "/intermediates/Earthworm_occurrence_GBIF-sWorm-Edapho-SoilReCon-JM.csv"),
          row.names = F)

# - - - - - - - - - - - - - - - - - - -
## Cleaning ####
# - - - - - - - - - - - - - - - - - - -

#setwd("D:/_students/Romy/SoilBiodiversity")

gc()
library(tidyverse)
library(here)

library(CoordinateCleaner)

Taxon_name <- "Crassiclitellata"
speciesNames <- read.csv(file=paste0("./results/Species_list_", Taxon_name, ".csv"))

data_raw <- read.csv(file=paste0(here::here(), "/intermediates/Earthworm_occurrence_GBIF-sWorm-Edapho-SoilReCon-JM.csv"))

# - - - - - - - - - - - - - - - - - - -
# create a table to see how many records get removed.
df_cleaning <- tibble::tibble(CleaningStep="merged_RawData", NumberRecords=nrow(data_raw))

# remove data without coordinates or taxa names
data <- data_raw[complete.cases(data_raw$longitude, data_raw$latitude, data_raw$species),] #nrow=83,260
data

df_cleaning <- df_cleaning %>% add_row(CleaningStep="merged_coordinates", NumberRecords=nrow(data))

# remove records outside of Europe
data <- data %>% filter(extent_Europe[1] <= longitude &  longitude <= extent_Europe[2]) %>% 
  filter(extent_Europe[3] <= latitude &  latitude <= extent_Europe[4])
data # nrow=78,055

df_cleaning <- df_cleaning %>% add_row(CleaningStep="merged_Europe", NumberRecords=nrow(data))

# remove species with only sp in name
data <- data %>% filter(!stringr::str_detect(data$species, "[[:blank:]]sp"))
data # nrow=78,036

df_cleaning <- df_cleaning %>% add_row(CleaningStep="merged_species", NumberRecords=nrow(data))
df_cleaning

data$OBJECTID <- 1:nrow(data) 

# - - - - - - - - - - - - - - - - - - -
## CoordinateCleaner ####
# flag problems with coordinates
dat_cl <- data.frame(data)
flags <- CoordinateCleaner::clean_coordinates(x = dat_cl, lon = "longitude", lat = "latitude",
                                              species = "species", tests = c("capitals", "centroids", "equal", "gbif", "zeros", "seas"), #normally: test "countries"
                                              country_ref = rnaturalearth::ne_countries("small"), 
                                              country_refcol = "iso_a3", urban_ref = NULL)
sum(flags$.summary) #those not flagged! = 75526 out of 78036

# save flagged coordinates
write.csv(flags %>% filter(!.summary), file=paste0(here::here(), "/results/FlaggedRecords_Crassiclitellata.csv"), row.names=F)

# remove flagged records from the clean data (i.e., only keep non-flagged ones)
dat_cl <- dat_cl[flags$.summary, ]

df_cleaning <- df_cleaning %>% add_row(CleaningStep="merged_CoordinateCleaner", NumberRecords=nrow(dat_cl))

## Plot flagged records
world.inp <- map_data("world")

pdf(paste0(here::here(), "/figures/CoordinateCleaner_flagged_records_Crassiclitellata.pdf"), width=20)
ggplot() + 
  geom_map(data = world.inp, map = world.inp, aes(map_id = region), fill = "grey80") + 
  xlim(min(data$longitude, na.rm = T), max(data$longitude, na.rm = T)) + 
  ylim(min(data$latitude, na.rm = T), max(data$latitude, na.rm = T)) + 
  geom_point(data = data_raw, aes(x = longitude, y = latitude), colour = "darkred", size = 1, show.legend = T) +
  geom_point(data = dat_cl, aes(x = longitude, y = latitude), colour = "darkgreen", size = 1, show.legend = T) + 
  coord_fixed() + 
  scale_color_manual(name='CoordinateCleaner',
                     values=c('RawRecords = red'='darkred', 'CleanRecords = green'='darkgreen'))+ 
  theme_bw() + theme(axis.title = element_blank())
dev.off()

# - - - - - - - - - - - - - - - - - - -
## Some structuring ####
# add species ID
dat_cl <- dplyr::full_join(dat_cl, speciesNames[,c("Species", "SpeciesID")], 
                            by=c("species" = "Species"))

# create x and y column
dat_cl <- dat_cl %>% mutate("x"=longitude, "y"=latitude)

# remove NA in coordinates (should not change nrow() as we filtered for complete.cases before)
dat_cl <- dat_cl[complete.cases(dat_cl$x),]
dat_cl <- dat_cl[complete.cases(dat_cl$y),]

# - - - - - - - - - - - - - - - - - - -
## Save clean data ####
write.csv(dat_cl, file=paste0(here::here(), "/intermediates/Occurrences_", Taxon_name, ".csv"),
          row.names = F)

# load number of records during cleaning process
df_cleaning1 <- read.csv(file=paste0(here::here(), "/results/NoRecords_cleaning_", Taxon_name, ".csv"))

df_cleaning <- df_cleaning1 %>% full_join(df_cleaning)
df_cleaning

# save updated number of records during cleaning process
write.csv(df_cleaning, file=paste0(here::here(), "/results/NoRecords_cleaning_", Taxon_name, ".csv"), row.names = F)
